// frontend/src/app/models/login.model.ts
export interface Credential {
  username: string;
  password: string;
}

export interface FirstAccessInfo {
  username: string;
  fullName: string;
  phone: string;
  email: string;
  cpf: string;
  rg: string;
  birthday: number;
  cep: string;
  city: string;
  uf: string;
  street: string;
  number: string;
  neighborhood: string;
  complement: string;
}